Added graph visualizer
We can now see what the graph network looks like. As well as the highlighted best path, found from the simulation. If we want to only view the network and not the best path, then there is a boolean flag option which can be set to false to disable to visualizer. This requires Pyvis python library. I installed the latest, 3.2. Also python's built-in 'webbrowser' module is used to open up the HTML network.